upgrade basic api

+++ be_camphub/src/main/java/org/camphub/be_camphub/dto/response/item/ItemResponse.java
@@ -19,8 +19,8 @@ public class ItemResponse {
     String categoryName;
     String name;
     String description;
-    BigDecimal price;
-    BigDecimal depositAmount;
+    Double price;
+    Double depositAmount;
     String status;
     List<String> imageUrls;
     LocalDateTime createdAt;
